Rename Server Certs tab to Server Identity with two-section layout
Restructure the tab into two always-visible sections:
1. Certificates (SPIFFE SVIDs) - existing cert table with SPIFFE ID, fingerprint, status
2. Server Keys (gSRK) - new section showing publicKey and signedPublicKey fields

Both sections share the same data source and approval/revoke actions.
The ServerCertEntry interface is extended with publicKey and signedPublicKey fields.
